Inline NotifyingUpdateFunctionRunner
