+++ subprojects/core/src/testFixtures/groovy/org/gradle/api/internal/file/TestFiles.java
@@ -40,7 +40,6 @@ import org.gradle.internal.resource.local.FileResourceConnector;
 import org.gradle.internal.resource.local.FileResourceRepository;
 import org.gradle.internal.snapshot.AtomicSnapshotHierarchyReference;
 import org.gradle.internal.snapshot.CaseSensitivity;
-import org.gradle.internal.snapshot.SnapshotHierarchy;
 import org.gradle.internal.time.Time;
 import org.gradle.internal.vfs.VirtualFileSystem;
 import org.gradle.internal.vfs.impl.DefaultSnapshotHierarchy;
@@ -195,12 +194,12 @@ public class TestFiles {
 
     public static VirtualFileSystem virtualFileSystem() {
         CaseSensitivity caseSensitivity = fileSystem().isCaseSensitive() ? CASE_SENSITIVE : CASE_INSENSITIVE;
-        AtomicSnapshotHierarchyReference root = new AtomicSnapshotHierarchyReference(DefaultSnapshotHierarchy.empty(caseSensitivity), SnapshotHierarchy.UpdateFunctionRunner.WITHOUT_LISTENERS);
+        AtomicSnapshotHierarchyReference rootReference = new AtomicSnapshotHierarchyReference(DefaultSnapshotHierarchy.empty(caseSensitivity), path -> true);
         return new DefaultVirtualFileSystem(
             fileHasher(),
             new StringInterner(),
             fileSystem(),
-            root,
+            rootReference,
             locations -> {}
         );
     }
