+++ subprojects/file-watching/src/test/groovy/org/gradle/internal/watch/vfs/impl/WatchingNotSupportedFileSystemWatchingHandlerTest.groovy
@@ -27,7 +27,7 @@ class WatchingNotSupportedFileSystemWatchingHandlerTest extends Specification {
     def nonEmptySnapshotHierarchy = Stub(SnapshotHierarchy) {
         empty() >> emptySnapshotHierarchy
     }
-    def root = new AtomicSnapshotHierarchyReference(nonEmptySnapshotHierarchy, SnapshotHierarchy.UpdateFunctionRunner.WITHOUT_LISTENERS)
+    def root = new AtomicSnapshotHierarchyReference(nonEmptySnapshotHierarchy, { path -> true })
     def watchingNotSupportedHandler = new WatchingNotSupportedFileSystemWatchingHandler(root)
 
     def "invalidates the virtual file system before and after the build"() {
